Moved mock data from tests
