+++ tests/MockData/Models/InvalidModel.php
@@ -2,6 +2,6 @@
 
 declare(strict_types=1);
 
-namespace AbstractRepo\Test\Models;
+namespace AbstractRepo\Test\MockData\Models;
 
 class InvalidModel{}
