+++ tests/MockData/Models/T4.php
@@ -2,19 +2,22 @@
 
 declare(strict_types=1);
 
-namespace AbstractRepo\Test\Models;
+namespace AbstractRepo\Test\MockData\Models;
 
 use AbstractRepo\Attributes\Entity;
+use AbstractRepo\Attributes\ForeignKey;
 use AbstractRepo\Attributes\PrimaryKey;
 use AbstractRepo\Attributes\Searchable;
+use AbstractRepo\Enums\Relationship;
 use AbstractRepo\Interfaces\IModel;
 
-#[Entity('T3')]
-class T3 implements IModel{
+#[Entity('T4')]
+class T4 implements IModel{
     public function __construct(
         #[PrimaryKey(false)]
+        #[ForeignKey(relationship: Relationship::MANY_TO_ONE, columnName: 't3_id')]
         #[Searchable]
-        public string $id,
+        public T3 $t3,
         #[Searchable]
         public string $v1
     ){}
