+++ tests/MockData/Repository/T1Repository.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace AbstractRepo\Test\Repository;
+namespace AbstractRepo\Test\MockData\Repository;
 
 use AbstractRepo\Repository\AbstractRepository;
-use AbstractRepo\Test\Models\T1;
+use AbstractRepo\Test\MockData\Models\T1;
 
 class T1Repository extends AbstractRepository
 {
