+++ tests/MockData/Repository/T3Repository.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace AbstractRepo\Test\Repository;
+namespace AbstractRepo\Test\MockData\Repository;
 
 use AbstractRepo\Repository\AbstractRepository;
-use AbstractRepo\Test\Models\T3;
+use AbstractRepo\Test\MockData\Models\T3;
 
 class T3Repository extends AbstractRepository
 {
