+++ tests/Repository/TestInvalidModelRepository.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace AbstractRepo\Test\Repository;
 
 use AbstractRepo\Repository\AbstractRepository;
-use AbstractRepo\Test\Models\InvalidModel;
+use AbstractRepo\Test\MockData\Models\InvalidModel;
 
 class TestInvalidModelRepository extends AbstractRepository
 {
